refactor: update import paths for ItemDTO to improve consistency; remove unused models
