+++ backend/models/item_dto.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import Optional
 
 from pydantic import BaseModel
 
@@ -28,24 +28,3 @@ def item_to_dto(item):
             rarity=getattr(item, "rarity", None),
             float_value=getattr(item, "float_value", None),
         )
-
-
-class ListingQueryParams(BaseModel):
-    cursor: Optional[str] = None
-    limit: Optional[int] = 50
-    sort_by: Optional[str] = "best_deal"
-    category: Optional[int] = 0
-    def_index: Optional[List[int]] = None
-    min_float: Optional[float] = None
-    max_float: Optional[float] = None
-    rarity: Optional[int] = None
-    paint_seed: Optional[List[int]] = None
-    paint_index: Optional[int] = None
-    user_id: Optional[str] = None
-    collection: Optional[str] = None
-    min_price: Optional[float] = None
-    max_price: Optional[float] = None
-    market_hash_name: Optional[str] = None
-    item_name: Optional[str] = None
-    type: Optional[str] = None
-    stickers: Optional[str] = None
